Add class method newGame
